+++ src/modules/translation/translation.service.ts
@@ -1,14 +1,86 @@
 import { Injectable } from '@nestjs/common';
 import { MealIngredient } from '../ingredient/ingredient.types';
 import { Language } from '../../common/types';
-import { TranslatedIngredient } from './translation.types';
-import { MealRecipeSection, MealRecipeSections } from '../meal/meal.types';
+import { TranslatedDetailedMeal, TranslatedIngredient } from './translation.types';
+import { DetailedMeal, MealRecipeSection, MealRecipeSections } from '../meal/meal.types';
 import { compoundTextToTranslate, convertAmountToText, normalizeName, normalizeUnit } from '../../common/helpers';
 import translate from '@iamtraction/google-translate';
 
 @Injectable()
 export class TranslationService {
 
+    async translateMeal(meal: DetailedMeal, targetLanguage?: Language): Promise<TranslatedDetailedMeal> {
+        // NOTE: Do not translate when your language is English
+        if (['en', 'en-US'].includes(targetLanguage)) {
+            return {
+                description: '',
+                ingredients: [],
+                recipe: []
+            };
+        }
+
+        const { description, ingredients, recipeSections } = meal;
+        const recipeNewSectionIndexes: number[] = recipeSections.reduce((acc, curr) => {
+            if (acc.length === 0) {
+                return [0];
+            }
+
+            return [...acc, acc.reduce((els, el) => els + el, 0) + curr.steps.length + 1];
+        }, []);
+        const ingredientImages: string[] = [];
+        const startRecipeIdx: number = ingredients.length;
+
+        const stringToTranslate: string = [
+            description,
+            ...ingredients.map(ingredient => {
+                const { amount, unit, name, imageUrl } = ingredient;
+                const normalizedName = normalizeName(name);
+                const normalizedUnit = normalizeUnit(amount, unit);
+                const textAmount = convertAmountToText(amount);
+                const compoundedText = compoundTextToTranslate(textAmount, normalizedUnit, normalizedName);
+
+                ingredientImages.push(imageUrl);
+
+                return compoundedText;
+            }),
+            ...recipeSections.map(section => {
+                return [
+                    section.name,
+                    ...section.steps
+                ];
+            }).flat()
+        ].join('\n');
+
+        const translatedResult = await this.translate(stringToTranslate, targetLanguage);
+        const [ translatedDescription, ...translatedIngredientsAndRecipes] = translatedResult.split('\n');
+        const translatedIngredients = translatedIngredientsAndRecipes.slice(0, startRecipeIdx);
+        const translatedRecipes = translatedIngredientsAndRecipes.slice(startRecipeIdx);
+
+        const ingredientList: TranslatedIngredient[] = translatedIngredients.map((ingredient, idx) => ({ text: ingredient, imageUrl: ingredientImages[idx] }));
+        const translatedRecipeSections: MealRecipeSections = recipeNewSectionIndexes.map((index, idx, sections) => {
+            const nextSection = sections[idx + 1];
+
+            if (!nextSection) {
+                return {
+                    name: translatedRecipes[index],
+                    steps: translatedRecipes.slice(index + 1)
+                };
+            }
+
+            return {
+                name: translatedRecipes[index],
+                steps: translatedRecipes.slice(index + 1, nextSection)
+            };
+        });
+
+        return {
+            description: translatedDescription,
+            ingredients: ingredientList,
+            recipe: translatedRecipeSections
+        };
+    }
+
+    // FIXME: Deprecated
     async translateDescription(description: string, targetLanguage?: Language): Promise<string> {
         // NOTE: Do not translate when your language is English
         if (['en', 'en-US'].includes(targetLanguage)) {
@@ -18,6 +90,7 @@ export class TranslationService {
         return await this.translate(description, targetLanguage);
     }
 
+    // FIXME: Deprecated
     async translateRecipe(recipeSections: MealRecipeSections, targetLanguage?: Language): Promise<MealRecipeSections> {
         // NOTE: Do not translate when your language is English
         if (['en', 'en-US'].includes(targetLanguage)) {
@@ -42,6 +115,7 @@ export class TranslationService {
         };
     }
 
+    // FIXME: Deprecated
     async translateIngredients(ingredients: MealIngredient[], targetLanguage?: Language): Promise<TranslatedIngredient[]> {
         // NOTE: Do not translate when your language is English
         if (['en', 'en-US'].includes(targetLanguage)) {
@@ -63,6 +137,10 @@ export class TranslationService {
     }
 
     async translate(text: string, targetLanguage: Language): Promise<string> {
+        if (text === '') {
+            return '';
+        }
+
         const language = this._getTargetLanguage(targetLanguage);
         const result = await translate(text, { from: 'en', to: language });